LayoutTests/http/tests/security/xssAuditor/xss-protection-parsing-02-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.

This tests that the X-XSS-Protection header is not ignored when there is a trailing semicolon. Although theoretically malformed, we tolerate this case without issuing an error.



--------
Frame: 'frame'
--------
If you see this message and no JavaScript alert() then the test PASSED.
